Major changes to the UI logic
Updates were made to how the UI is instacniated so that keyEvents could properly function.

Current activated widget function has been created to track what widget is in use.

Table range copy keyEvent operation added. Checks if current widget is instace of QTableWidget.
If so it will copy the current selected range to the clipboard. This enables the functionality
of copying data out of the TableWidget